bible-read-api4

+++ qtai-server/src/main/java/com/qtai/domain/bible/api/dto/BibleBookResponse.java
@@ -1,11 +1,11 @@
 package com.qtai.domain.bible.api.dto;
 
 public record BibleBookResponse(
-        Short id,
+        Integer id,
         String testament,
         String code,
         String koreanName,
         String englishName,
-        Short displayOrder
+        Integer displayOrder
 ) {
 }
